refactor(config): 攤平 top-level façade re-export 至 shared canonical
延續 #487(env.js 攤平)方向,移除 config-v3 過渡期遺留的 top-level re-export façade,
caller 直接導入 canonical source。

- 刪除 4 個 façade: icons.js / highlightConstants.js / notionCodeLanguages.js / saveStatus.js
- 15 個 caller(6 scripts/ + 9 pages/)改為直接導入 config/shared/*.js
- config/index.js 移除冗餘 re-export(shared/index.js 已涵蓋,barrel export 集合不變)
- 同步移除 icons/highlightConstants 的 coverage exclusion(jest/sonar/契約 test 三方對齊)

驗證: config 123 tests + 受影響模組 3366 tests 全綠; lint clean; 無殘留 façade 引用
