+++ pages/options/MigrationTool.js
@@ -3,8 +3,7 @@
  * 負責舊版標註數據的遷移 UI 與協調
  */
 /* global chrome */
-import { UI_ICONS } from '../../scripts/config/icons.js';
-import { COMMON_CSS_CLASSES } from '../../scripts/config/shared/ui.js';
+import { UI_ICONS, COMMON_CSS_CLASSES } from '../../scripts/config/shared/ui.js';
 import { RUNTIME_ACTIONS } from '../../scripts/config/shared/runtimeActions.js';
 import { ERROR_MESSAGES, UI_MESSAGES } from '../../scripts/config/shared/messages.js';
 import Logger from '../../scripts/utils/Logger.js';
