+++ scripts/background/handlers/saveHandlers.js
@@ -26,7 +26,7 @@ import { ErrorHandler } from '../../utils/ErrorHandler.js';
 import { CONTENT_QUALITY } from '../../config/shared/content.js';
 import { ERROR_MESSAGES } from '../../config/messages/errorMessages.js';
 import { RUNTIME_ACTIONS } from '../../config/shared/runtimeActions.js';
-import { SAVE_STATUS_KINDS, createSaveStatusResponse } from '../../config/saveStatus.js';
+import { SAVE_STATUS_KINDS, createSaveStatusResponse } from '../../config/shared/saveStatus.js';
 import { isRestrictedInjectionUrl } from '../services/InjectionService.js';
 import { resolveSaveStatus } from '../services/SaveStatusCoordinator.js';
 import {
