+++ scripts/background/services/SaveStatusCoordinator.js
@@ -1,5 +1,5 @@
 import { HANDLER_CONSTANTS } from '../../config/shared/core.js';
-import { SAVE_STATUS_KINDS, createSaveStatusResponse } from '../../config/saveStatus.js';
+import { SAVE_STATUS_KINDS, createSaveStatusResponse } from '../../config/shared/saveStatus.js';
 
 function defaultLogger() {
   return {
