--- scripts/config/highlightConstants.js
@@ -1,1 +1,0 @@
-export * from './shared/highlightConstants.js';
