--- scripts/config/icons.js
@@ -1,1 +1,0 @@
-export { UI_ICONS } from './shared/ui.js';
