--- scripts/config/notionCodeLanguages.js
@@ -1,1 +1,0 @@
-export * from './shared/notionCodeLanguages.js';
