+++ sonar-project.properties
@@ -13,7 +13,7 @@ sonar.exclusions=node_modules/**,dist/**,coverage/**,tests/**,**/*.test.js,**/*.
 # 覆蓋率排除清單（對齊 jest.config.js 的 collectCoverageFrom）
 # 兩類排除：(1) 純常量 config；(2) prod DCE 後 unreachable 的 Toolbar 鏈（見 jest.config.js 同段註釋）。
 # 這些檔案的覆蓋率不代表 production 行為，但仍接受 SonarCloud 的 issue / code smell 分析。
-sonar.coverage.exclusions=scripts/config/app.js,scripts/config/icons.js,scripts/config/ui.js,scripts/config/extraction.js,scripts/config/highlightConstants.js,scripts/config/index.js,scripts/config/extension/**/*.js,scripts/highlighter/ui/Toolbar.js,scripts/highlighter/ui/ToolbarRuntime.js,scripts/highlighter/ui/ToolbarState.js,scripts/highlighter/ui/ToolbarUI.js,scripts/highlighter/ui/styles/toolbarStyles.js,scripts/highlighter/ui/components/ColorPicker.js,scripts/highlighter/ui/components/MiniIcon.js,scripts/highlighter/ui/components/ToolbarContainer.js,**/*.test.js,**/*.spec.js,tests/**,node_modules/**,dist/**,coverage/**,playwright-report/**,lib/Readability.js
+sonar.coverage.exclusions=scripts/config/app.js,scripts/config/ui.js,scripts/config/extraction.js,scripts/config/index.js,scripts/config/extension/**/*.js,scripts/highlighter/ui/Toolbar.js,scripts/highlighter/ui/ToolbarRuntime.js,scripts/highlighter/ui/ToolbarState.js,scripts/highlighter/ui/ToolbarUI.js,scripts/highlighter/ui/styles/toolbarStyles.js,scripts/highlighter/ui/components/ColorPicker.js,scripts/highlighter/ui/components/MiniIcon.js,scripts/highlighter/ui/components/ToolbarContainer.js,**/*.test.js,**/*.spec.js,tests/**,node_modules/**,dist/**,coverage/**,playwright-report/**,lib/Readability.js
 
 # 測試覆蓋率報告路徑（jest unit only；E2E coverage 不合併進 quality gate，避免 ghost-line 污染）
 sonar.javascript.lcov.reportPaths=coverage/jest/lcov.info
